[DEV-9731] create Parent component
