+++ packages/chart/src/index.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
 
-import CdcChart from './CdcChart'
+import CdcChart from './CdcChartComponent'
 import './coreStyles_chart.scss'
 
 import 'react-tooltip/dist/react-tooltip.css'
