+++ packages/dashboard/src/components/VisualizationRow.tsx
@@ -3,7 +3,7 @@ import React, { useContext, useMemo } from 'react'
 import Toggle from './Toggle'
 import _ from 'lodash'
 import { ConfigRow } from '../types/ConfigRow'
-import CdcChart from '@cdc/chart/src/CdcChart'
+import CdcChartComponent from '@cdc/chart/src/CdcChartComponent'
 import CdcDataBite from '@cdc/data-bite/src/CdcDataBite'
 import CdcMap from '@cdc/map/src/CdcMap'
 import CdcWaffleChart from '@cdc/waffle-chart/src/CdcWaffleChart'
@@ -179,7 +179,7 @@ const VisualizationRow: React.FC<VizRowProps> = ({
                 row={row}
               >
                 {visualizationConfig.type === 'chart' && (
-                  <CdcChart
+                  <CdcChartComponent
                     key={col.widget}
                     config={visualizationConfig}
                     dashboardConfig={config}
